Fix Android build
For some reason this line of code fails on Android with a missing Copy
trait error, but not on all other platforms. Different library versions
could explain why the Copy trait is not implemented on Android, but
apparently everywhere else.

Anyhow, with this change the code should be correct on all platforms.